requirement.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
decorator==4.0.9
Django==1.8
django-ckeditor==5.0.3
django-haystack==2.4.1
django-json-response==1.1.3
django-markdown==0.8.4
django-pagedown==0.1.1
ipython==4.2.0
ipython-genutils==0.1.0
isort==4.2.5
jieba==0.38
Markdown==2.6.6
markdown2==2.3.1
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
traitlets==4.2.1
uWSGI==2.0.12
Whoosh==2.7.4
